# openapi.yaml
openapi: 3.1.1
paths:
...
  /:
    post:
      operationId: UrlsController_shortenUrl
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LongUrlDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortenedUrlDto'
      tags:
        - Shortened URLs
  /{short}:
    get:
      operationId: UrlsController_visitShortUrl
      parameters:
        - name: short
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LongUrlDto'
      tags:
        - Shortened URLs
...
info:
  title: Zero Width Shortener
  description: A URL shortener that uses zero width characters to shorten URLs.
  version: 2.0.0
  contact:
    name: Jonah Snider
    url: https://jonahsnider.com
    email: jonah@jonahsnider.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: https://zws.im/api
components:
  schemas:
...
    LongUrlDto:
      type: object
      properties:
        url:
          type: string
          maxLength: 500
          format: uri
      required:
        - url
      title: LongUrl
    ShortenedUrlDto:
      type: object
      properties:
        short:
          type: string
        url:
          type: string
          format: uri
      required:
        - short
        - url
      title: ShortenedUrl
...